LPD-71796 portal-search-api: rename package
https://liferay.atlassian.net/browse/LPD-71796
