+++ modules/apps/portal-search/portal-search-api/src/test/java/com/liferay/portal/search/ml/embedding/text/helper/TextEmbeddingContentHelperTest.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.portal.search.ml.embedding.text.util;
+package com.liferay.portal.search.ml.embedding.text.helper;
 
 import com.liferay.petra.string.StringBundler;
 import com.liferay.petra.string.StringPool;
